+++ HISTORY.md
@@ -6,6 +6,7 @@
 * Display observation count and leaf taxon count on taxon info cards
 * In refresh mode, check for taxonomy changes and update tags with the new taxon (1:1 changes only)
 * Add support for alternate XMP sidecar path format, if it already exists (`basename.ext.xmp` instead of `basename.xmp`)
+* Add CLI support for selecting a sidecar file directly (instead of via an associated image file)
 
 ## 0.7.0 (2022-07-29)
 * Rebuilt UI from scratch using Qt